# Optimize the buffer size in inverse_transform_add